feature
front

+++ api/db/migrate/20240227000541_create_financas.rb
@@ -3,7 +3,7 @@ class CreateFinancas < ActiveRecord::Migration[7.1]
     create_table :financas do |t|
       t.string :descricao
       t.integer :tipo_entrada
-      t.decimal :valor, precision: 8, scale: 2
+      t.float :valor, precision: 8, scale: 2
 
       t.timestamps
     end
